requirements.txt
PyYaml
termcolor
prettytable
requests
